chore: Split S3 types and impl blocks
